Moved getManifest to Manifest.
