+++ src/Cacher.ts
@@ -1,8 +1,4 @@
-import Manifest from "./Manifest";
-
 interface Cacher {
-    getManifest(manifestUrl: URL): Promise<Manifest>;   
-
     // This should kick off the downloads to enable offline use.
     enable(): Promise<void>;
