+++ src/Manifest.ts
@@ -1,3 +1,5 @@
+import Store from "./Store";
+
 export interface Metadata {
     title?: string;
     author?: string;
@@ -20,6 +22,29 @@ export default class Manifest {
     public readonly resources: Array<Link>;
     public readonly toc: Array<Link>;
     private readonly manifestUrl: URL;
+
+    public static async getManifest(manifestUrl: URL, store?: Store): Promise<Manifest> {
+        try {
+            const response = await window.fetch(manifestUrl.href)
+            const manifestJSON = await response.json();
+            if (store) {
+                await store.set("manifest", JSON.stringify(manifestJSON));
+            }
+            return new Manifest(manifestJSON, manifestUrl);
+        } catch (err) {
+            // We couldn't fetch the response, but there might be a cached version.
+            if (store) {
+                const manifestString = await store.get("manifest");
+                if (manifestString) {
+                    const manifestJSON = JSON.parse(manifestString);
+                    return new Manifest(manifestJSON, manifestUrl);
+                }
+            }
+
+            // There's nothing we can do, reraise the error.
+            throw err;
+        }
+    }
 
     public constructor(manifestJSON: any, manifestUrl: URL) {
         this.metadata = manifestJSON.metadata || {};
